Fix checkbox behaviour in IE
In IE checkboxes react to "change" event differently. In fact if you use
preventDefault() the checkbox behaves strangely. Form uses change events
to track form changes, and it prevents default behavor of onChange if
new value of the field matches the default value.

It's possible that there is still isue with that change event firing
on IE, but at least with this fix checkboxes are behaving better on IE

I'm going to create a test-case to finalize this fix