Update lib/Model/Table.php
Important patch. Model_Table saves empty string values as "" and not as NULL.
In most of the cases it's OK, but problems starts when you try to save empty DATE or DATETIME values.
Then MySQL saves them as some default past date (1970-01-01 in my case on Windows).

This patch fixes this problem and I guess it can't do any bad. I'm especially checking for "0" values which should be stored as 0 and not NULL.