Tricky bug
if you have Model, which has boolean field, but no default value
and if you try to set the value to false and save, it was ommitting
that, because exising NULL was casted to string as was the false,
making improper assumption that both are same. moreover, in this case
it was database settings that determined the value to be inserted
which could be "Y", even if you set the field value to false

Fixed by adding strong comparrison between existing value and the
requested value. false is not the same as null.