Model->set(): fix for array controller
If we use array controller for model where data source is array, then inserting NULL values for fields don't create appropriate array key.
That's because $value === $this->data[$name] if $this->data[$name] silently fails and as result is cast to NULL too. So NULL === NULL and field value equal NULL is not set (and as result - not saved).