CSS: upgraded reset.less to the one used by Bootstrap
Why? Because old reset.less did some bad things. For example, strictly
set font on absolutely all HTML tags. That way it was almost impossible
to use FontAwesome, for example.
I hope this reset will not break anything, but I would appreciate if you
could test it in some of your ready-to-use projects to see for yourself.