app->dbConnect to return the DB object and not API.
It’s unlikely that the dbConnect is ever used for chaining. Also
ability to specify custom DSN could be useful for executing the method
several times, when multiple connections should be maintained.

Overall - dbConnect will have to be extended for lazy-loaded of
multiple database connections as it might be required by add-ons.